Fixing FK error: meetings - meeting_timeslots

+++ database/migrations/2015_12_10_181512_create_meetings_table.php
@@ -18,7 +18,7 @@ class CreateMeetingsTable extends Migration
             $table->string('description'); //TODO improvement change with text? (remember to change request)
             $table->integer('group_id')->unsigned();
             $table->foreign('group_id')->references('id')->on('groups')->onDelete('cascade')->onUpdate('cascade');
-            $table->integer('start_time')->unsigned()->nullabe()->default(NULL);
+            $table->bigInteger('start_time')->unsigned()->nullabe()->default(NULL);
             $table->foreign('start_time')->references('id')->on('meeting_timeslots')->onDelete('cascade')->onUpdate('cascade');
             $table->integer('duration'); //minutes
             $table->timestamps();
